pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["IsacJfs <isac.jfs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.3"
factory-boy = "^3.3.0"
django = "^5.0"
django-rest-framework = "^0.1.0"
pytest-django = "^4.7.0"
django-extensions = "^3.2.3"
python-dotenv = "^1.0.0"
django-debug-toolbar = "^4.2.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
